chore: moved code from file, and documented range ptr function
feat: added function to get operand by its pointer
